shared/pyproject.toml
[project]
name = "aial-shared"
version = "0.1.0"
description = "AIAL shared models and utilities"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "PyJWT[crypto]>=2.8",
    "fastapi>=0.115",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aial_shared"]
